Cambios solicitados realizados. Se agregaron las clases UserController, IUserService y UserServiceImpl
