+++ src/main/java/com/alkemy/ong/auth/controller/AuthController.java
@@ -8,11 +8,9 @@ import com.alkemy.ong.auth.service.JwtUtils;
 import com.alkemy.ong.auth.service.impl.UserDetailsCustomService;
 import com.alkemy.ong.exception.BadRequestException;
 import com.alkemy.ong.exception.ForbiddenException;
-import com.alkemy.ong.model.UserEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -62,21 +60,6 @@ public class AuthController {
         }
         final String jwt = jwtTokenUtil.generateToken(userDetails);
         return ResponseEntity.ok(new AuthenticationResponse(jwt));
-    }
-
-    @GetMapping("/users")
-    public ResponseEntity<List<UserResponseDto>> getAllUsers() {
-        List<UserResponseDto> usersDTOs = this.userDetailsCustomService.getAllUsers();
-        return ResponseEntity.ok().body(usersDTOs);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable(name = "id") Long userId) {
-
-        this.userDetailsCustomService.deleteUserById(userId);
-
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-
     }
 
 }
